slitley modify upload format. about div and form tag location.

+++ WebContent/view.jsp
@@ -1,5 +1,4 @@
-<%@ page language="java" contentType="text/html; charset=UTF-8"
-	pageEncoding="UTF-8"%>
+<%@ page language="java" contentType="text/html; charset=UTF-8" pageEncoding="UTF-8"%>
 
 <%@ page import="java.io.PrintWriter"%>
 <%@ page import="bbs.Bbs"%>
@@ -20,21 +19,21 @@
 <body>
 	<%
 		String userID = null;
-		if(session.getAttribute("userID") != null){
-			userID = (String)session.getAttribute("userID");
+		if (session.getAttribute("userID") != null) {
+			userID = (String) session.getAttribute("userID");
 		}
-		
+
 		int bbsID = 0;
-		if(request.getParameter("bbsID") != null){
+		if (request.getParameter("bbsID") != null) {
 			bbsID = Integer.parseInt(request.getParameter("bbsID"));
 		}
-		
-		if(bbsID == 0){
+
+		if (bbsID == 0) {
 			PrintWriter script = response.getWriter();
 			script.println("<script>");
 			script.println("alert('유효하지 않은 글입니다.')");
 			script.println("location.href = 'bbs.jsp'");
-			script.println("</script>"); 
+			script.println("</script>");
 		}
 		Bbs bbs = new BbsDAO().getBbs(bbsID);
 	%>
@@ -44,31 +43,32 @@
 
 	<div class="container">
 		<div class="row">
-			<table class="table table-striped"
-				style="text-align: center; border: 1px solid #dddddd; word-break: break-all;">
+			<table class="table table-striped" style="text-align: center; border: 1px solid #dddddd; word-break: break-all;">
 				<thead>
 					<tr>
-						<th colspan="4"
-							style="background-color: #eeeeee; text-align: center;">게시판 글</th>
+						<th colspan="4" style="background-color: #eeeeee; text-align: center;">게시판 글</th>
 					</tr>
 				</thead>
 
 				<tbody>
 					<tr>
 						<td style="width: 20%;">글제목</td>
-						<td colspan="3"><%= bbs.getBbsTitle().replaceAll(" ", "&nbsp;").replaceAll("<", "&lt;").replaceAll(">","&gt;").replaceAll("\n", "<br>") %></td>
+						<td colspan="3"><%=bbs.getBbsTitle().replaceAll(" ", "&nbsp;").replaceAll("<", "&lt;").replaceAll(">", "&gt;")
+					.replaceAll("\n", "<br>")%></td>
 					</tr>
 					<tr>
 						<td>작성자</td>
-						<td colspan="3"><%= bbs.getUserID() %></td>
+						<td colspan="3"><%=bbs.getUserID()%></td>
 					</tr>
 					<tr>
 						<td>작성일자</td>
-						<td colspan="3"><%= bbs.getBbsDate().substring(0,11)+bbs.getBbsDate().substring(11,13)+"시"+bbs.getBbsDate().substring(14,16)+"분" %></td>
+						<td colspan="3"><%=bbs.getBbsDate().substring(0, 11) + bbs.getBbsDate().substring(11, 13) + "시"
+					+ bbs.getBbsDate().substring(14, 16) + "분"%></td>
 					</tr>
 					<tr>
 						<td>내용</td>
-						<td colspan="3" style="min-height: 200px; text-align: left;"><%=bbs.getBbsContent().replaceAll(" ", "&nbsp;").replaceAll("<", "&lt;").replaceAll(">","&gt;").replaceAll("\n", "<br>")%></td>
+						<td colspan="3" style="min-height: 200px; text-align: left;"><%=bbs.getBbsContent().replaceAll(" ", "&nbsp;").replaceAll("<", "&lt;").replaceAll(">", "&gt;")
+					.replaceAll("\n", "<br>")%></td>
 					</tr>
 				</tbody>
 			</table>
@@ -76,34 +76,35 @@
 			<table class="table table-striped">
 				<tbody>
 					<%
-									ReplyDAO replyDAO = new ReplyDAO();
-									ArrayList<Reply> list = replyDAO.getList(bbsID);
-				
-									for(int i=0; i<list.size(); i++){
-							%>
+						ReplyDAO replyDAO = new ReplyDAO();
+						ArrayList<Reply> list = replyDAO.getList(bbsID);
+
+						for (int i = 0; i < list.size(); i++) {
+					%>
 					<tr>
-						<td align="center"><%= list.get(i).getReplyID() %></td>
-						<td align="left" style="word-break: break-all;"><%= list.get(i).getReplyContent().replaceAll(" ", "&nbsp;").replaceAll("<", "&lt;").replaceAll(">","&gt;").replaceAll("\n", "<br>") %></td>
+						<td align="center"><%=list.get(i).getReplyID()%></td>
+						<td align="left" style="word-break: break-all;"><%=list.get(i).getReplyContent().replaceAll(" ", "&nbsp;").replaceAll("<", "&lt;")
+						.replaceAll(">", "&gt;").replaceAll("\n", "<br>")%></td>
 						<%
-										if(userID!=null && userID.equals(list.get(i).getUserID())){
-									%>
-						<td align="center"><a
-							onclick="return confirm('정말로 삭제하시겠습니까?')" a
-							href="replyDeleteAction.jsp?bbsID=<%=bbsID%>&replyID=<%=list.get(i).getReplyID() %>"
-							type="button" class="close" aria-label="close"> <span
-								aria-hidden="true">&times;</span>
-						</a></td>
+							if (userID != null && userID.equals(list.get(i).getUserID())) {
+						%>
+						<td align="center">
+							<a onclick="return confirm('정말로 삭제하시겠습니까?')" href="replyDeleteAction.jsp?bbsID=<%=bbsID%>&replyID=<%=list.get(i).getReplyID()%>" type="button" class="close" aria-label="close"> <span aria-hidden="true">&times;</span></a>
+						</td>
 						<%
-											}
-									%>
-						<td style="width: 10%;"><%= list.get(i).getUserID() %></td>
-						<td style="width: 15%;"><%= list.get(i).getReplyDate().substring(0,11)+list.get(i).getReplyDate().substring(11,13)+"시"+list.get(i).getReplyDate().substring(14,16)+"분" %></td>
+							}
+						%>
+						<td style="width: 10%;"><%=list.get(i).getUserID()%></td>
+						<td style="width: 15%;"><%=list.get(i).getReplyDate().substring(0, 11) + list.get(i).getReplyDate().substring(11, 13)
+						+ "시" + list.get(i).getReplyDate().substring(14, 16) + "분"%></td>
 					</tr>
 					<%
-								}
-							%>
+						}
+					%>
 					<tr>
-						<td><a href="fileDownload.jsp">파일 다운로드 페이지</a></td>
+						<td>
+							<a href="fileDownload.jsp">파일 다운로드 페이지</a>
+						</td>
 					</tr>
 					<tr>
 						<td>썸네일</td>
@@ -112,32 +113,40 @@
 			</table>
 		</div>
 
-		<table class="table table-condensed">
-			<form method="post" action="replyAction.jsp">
-				<tbody>
-					<td style="width: 90%;"><input type="text"
-						class="form-control" palceholder="댓글" name="replyContent"
-						maxlength="2048" style="height: 150px;"></td>
-					<td style="width: 10%; vertical-align: bottom;" align="center">
-						<input type="submit" class="btn btn-primary pull-right"
-						value="댓글작성">
-					</td>
-					<input type="hidden" name="bbsID" value="<%=bbs.getBbsID()%>">
-				</tbody>
-			</form>
-		</table>
+		<form method="post" action="replyAction.jsp">
+			<table class="table table-condensed">
 
+				<tbody>
+					<tr>
+						<td style="width: 90%;">
+							<input type="text" class="form-control" placeholder="댓글" name="replyContent" maxlength="2048" style="height: 150px;">
+						</td>
+					</tr>
+					<tr>
+						<td style="width: 10%; vertical-align: bottom;" align="center">
+							<input type="submit" class="btn btn-primary pull-right" value="댓글작성">
+						</td>
+					</tr>
+					<tr>
+						<td>
+							<input type="hidden" name="bbsID" value="<%=bbs.getBbsID()%>">
+						</td>
+					</tr>
+				</tbody>
+
+			</table>
+		</form>
+		
+		
 		<a href="bbs.jsp" class="btn btn-primary">목록</a>
 		<%
-					if(userID!=null && userID.equals(bbs.getUserID())){
-				%>
-		<a href="update.jsp?bbsID=<%=bbsID%>" class="btn btn-priamry">수정</a> <a
-			onclick="return confirm('정말로 삭제하시겠습니까?')"
-			href="deleteAction.jsp?bbsID=<%=bbsID%>" class="btn btn-priamry">삭제</a>
+			if (userID != null && userID.equals(bbs.getUserID())) {
+		%>
+		<a href="update.jsp?bbsID=<%=bbsID%>" class="btn btn-priamry">수정</a> <a onclick="return confirm('정말로 삭제하시겠습니까?')" href="deleteAction.jsp?bbsID=<%=bbsID%>" class="btn btn-priamry">삭제</a>
 		<%
-					}
-				%>
-		<input type="submit" class="btn btn-primary pull-right" value="글쓰기">
+			}
+		%>
+		<a href="write.jsp" class="btn btn-primary pull-right">글쓰기</a>
 
 	</div>
